feat: 4주차 멘토님 피드백 반영

+++ src/main/java/gift/administrator/option/OptionApiController.java
@@ -1,5 +1,6 @@
 package gift.administrator.option;
 
+import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.http.HttpStatus;
@@ -7,6 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -33,9 +36,21 @@ public class OptionApiController {
 
     @DeleteMapping("/options/{optionId}")
     public ResponseEntity<OptionDTO> deleteOptionByOptionId(
-        @PathVariable("optionId") Long optionId)
+        @PathVariable("optionId") long optionId)
         throws NotFoundException {
+        if (optionService.countAllOptionsByProductIdFromOptionId(optionId) == 1) {
+            throw new IllegalArgumentException("상품에는 적어도 하나의 옵션이 있어야 합니다.");
+        }
         optionService.deleteOptionByOptionId(optionId);
         return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @PutMapping("/products/{productId}/options/{optionId}")
+    public ResponseEntity<?> updateOptionByProductIdAndOptionId(
+        @PathVariable("productId") long productId, @PathVariable("optionId") long optionId,
+        @Valid @RequestBody OptionDTO optionDTO)
+        throws NotFoundException {
+        optionService.existsByNameSameProductIdNotOptionId(optionDTO.getName(), productId, optionId);
+        return ResponseEntity.ok(optionService.updateOption(optionId, optionDTO));
     }
 }
